inline small hook logic
